created a show-menu.rb
